refactor: move probot commands into `src/bin/` folder
