+++ src/bin/probot-run.ts
@@ -1,0 +1,3 @@
+import { Probot } from "../";
+
+Probot.run(process.argv);
